fw_att_control rename class file
